updated:updateMovie to fetch and push Casting and added rxjava to merge api between database

+++ app/src/main/java/com/example/enes/cinemaapp/service/Service.java
@@ -1,17 +1,10 @@
 package com.example.enes.cinemaapp.service;
-
-import com.example.enes.cinemaapp.data.model.Cast;
-import com.example.enes.cinemaapp.data.model.CastGetting;
 import com.example.enes.cinemaapp.data.model.Movie;
 import com.example.enes.cinemaapp.data.model.MovieGetting;
-
-
-import io.reactivex.Observable;
 import io.reactivex.Single;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
-
 
 public interface Service {
 
@@ -19,5 +12,5 @@ public interface Service {
     Single<MovieGetting<Movie>> getPopularMovies(@Query("page") Integer pageNo);
 
     @GET("movie/{movie_id}")
-    Single<Movie> getMovieCredits(@Path("movie_id") long movieId, @Query("append_to_response") String credits);
+    Single<Movie> getMovieCredits(@Path("movie_id") Integer movieId, @Query("append_to_response") String credits);
 }
